added new operations into input!

+++ src/maths/inc/maths/maths.h
@@ -11,14 +11,14 @@ extern double mul(double a, double b);
 extern double div(double a, double b);
 // divide reminder a and b
 extern double divr(double a, double b);
-extern long long factorial(long long a);
+extern unsigned long long factorial(unsigned long long a);
 
 extern double pow(double a, unsigned long long n);
 
 extern double sqrt(double a);
 extern double sqrtn(double n, unsigned long long a);
 
-extern long long GCD(long long a, long long b);
+extern double GCD(double a, double b);
 
 extern double log(double a, double n);
 extern double log2(double a);
